DPNG-8765 Implemented AuditTrail functionality

+++ models/audit_trail.go
@@ -1,8 +1,7 @@
 package models
 
 type AuditTrail struct {
-	//todo change int to time
-	CreatedOn     int    `json:"createdOn"`
+	CreatedOn     int64  `json:"createdOn"`
 	CreatedBy     string `json:"createdBy"`
-	LastUpdatedOn int    `json:"lastUpdatedOn"`
+	LastUpdatedOn int64  `json:"lastUpdatedOn"`
 }
